refactor(designsystem): 生成コードではないのでcolor_scheme.dartに再命名する
